Feature: 1. ASIN first crawler

+++ bin/release/deploy.sh
@@ -22,27 +22,23 @@ bin=$(cd "$bin">/dev/null || exit; pwd)
 APP_HOME=$(cd "$bin"/..>/dev/null || exit; pwd)
 
 echo "Deploy the project ..."
-echo "Changing version ..."
 
 SNAPSHOT_VERSION=$(head -n 1 "$APP_HOME/VERSION")
-VERSION=${SNAPSHOT_VERSION//"-SNAPSHOT"/""}
-echo "$VERSION" > "$APP_HOME"/VERSION
-find "$APP_HOME" -name 'pom.xml' -exec sed -i "s/$SNAPSHOT_VERSION/$VERSION/" {} \;
+VERSION="$SNAPSHOT_VERSION"
 
-mvn clean
-mvn package
+#echo "Changing version ..."
+#VERSION=${SNAPSHOT_VERSION//"-SNAPSHOT"/""}
+#echo "$VERSION" > "$APP_HOME"/VERSION
+#find "$APP_HOME" -name 'pom.xml' -exec sed -i "s/$SNAPSHOT_VERSION/$VERSION/" {} \;
+
+#mvn clean
+#mvn package
 
 ######################################################
 # Deploy
 ######################################################
 
-SOURCE="$APP_HOME/target/exotic-amazon-$VERSION.jar"
-if [ ! -e "$SOURCE" ]; then
-  echo "$SOURCE does not exist"
-  exit 1
-fi
-PACK_DIR=target/package
-
+PACK_DIR="$APP_HOME/target/exotic-amazon-$VERSION"
 if [ -e "$PACK_DIR" ]; then
   rm -r "$PACK_DIR"
 fi
@@ -51,14 +47,26 @@ mkdir -p "$PACK_DIR"
 mkdir "$PACK_DIR"/bin
 mkdir "$PACK_DIR"/conf
 
-cp "$SOURCE" "$PACK_DIR"
-cp VERSION "$PACK_DIR"
-cp -r "$APP_HOME"/bin "$PACK_DIR"
+JAR="$APP_HOME/target/exotic-amazon-$VERSION.jar"
+if [ ! -e "$JAR" ]; then
+  echo "$JAR does not exist"
+  exit 1
+fi
+
+cp "$JAR" "$PACK_DIR/"
+cp VERSION "$PACK_DIR/"
+cp -r "$APP_HOME"/bin "$PACK_DIR/"
 cp src/main/resources/config/* "$PACK_DIR/conf/"
 cp src/main/resources/logback-prod.xml "$PACK_DIR/conf/"
 
-ARCHIVE="target/exotic-amazon-$VERSION.tar.gz"
-tar -czf "$ARCHIVE" "$PACK_DIR"
+ARCHIVE_FILE="exotic-amazon-$VERSION.tar.gz"
+cd "$APP_HOME/target" || exit
+echo "Make tarball in directory (should be target): "
+pwd
+tar -czf "$ARCHIVE_FILE" "exotic-amazon-$VERSION"
+cd - || exit
+echo "Return to application home: "
+pwd
 
 if [[ -z "$TARGET_HOST" ]]; then
   echo "Target host not specified"
@@ -74,7 +82,7 @@ DESTINATION_BASE_DIR="/home/$REMOTE_USER/wwwpub/pub/exotic/exotic-amazon"
 DESTINATION_DIR="$DESTINATION_BASE_DIR/$VERSION"
 DESTINATION="$REMOTE_USER@$TARGET_HOST:$DESTINATION_DIR/"
 
-rsync --update -raz --progress "$ARCHIVE" "$DESTINATION"
+rsync --update -raz --progress "$APP_HOME/target/$ARCHIVE_FILE" "$DESTINATION"
 
 echo "Finished at" "$(date)"
